+++ src/server/auth.ts
@@ -19,8 +19,9 @@ import {
 } from "~/utils/auth/jwt";
 import { login } from "~/services/auth.service";
 
-import { User } from "@prisma/client";
-import { LoginSchema } from "./schema/zod-schema";
+import { User, Role } from "@prisma/client";
+import { LoginSchema } from "~/zod/authZ";
+
 
 /**
  * Module augmentation for `next-auth` types. Allows us to add custom properties to the `session`
@@ -36,7 +37,7 @@ declare module "next-auth" {
     name?: string | null;
     email?: string | null;
     image?: string | null;
-    role?: string | null;
+    role?: Role | null;
   }
 
   interface AdapterUser {
@@ -46,14 +47,14 @@ declare module "next-auth" {
     name?: string | null;
     email?: string | null;
     image?: string | null;
-    role?: string | null;
+    role?: Role | null;
   }
 
   interface Session extends DefaultSession {
     user: DefaultSession["user"] & {
       id: string;
       // ...other properties
-      role: string;
+      role: Role;
     };
     accessToken: string;
   }
@@ -70,7 +71,7 @@ declare module "next-auth/jwt" {
   interface JWT {
     iat: number;
     exp: number;
-    role: string; 
+    role: Role; 
     accessToken: string;
     refreshToken: string;
   }
@@ -94,7 +95,7 @@ export const authOptions: NextAuthOptions = {
           sub: user.id,
           name: user.name,
           email: user.email,
-          role: user.role,
+          role: user.role!,
           accessToken: user.accessToken,
           refreshToken: user.refreshToken,
           iat: Math.floor(Date.now() / 1000),
@@ -162,7 +163,7 @@ export const authOptions: NextAuthOptions = {
         session.user.id = token.sub;
         session.user.name = token.name;
         session.user.email = token.email!;
-        session.user.role = token.role!;
+        session.user.role = token.role!; 
         session.accessToken = token.accessToken;
       }
