+++ src/zod/attendenceZ.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+
+const markTeamAttendanceSchema = z.object({
+    teamId: z.string(),
+    eventId: z.string(),
+})
+
+export {
+    markTeamAttendanceSchema
+
+}
